+++ frontend/pages/_app.tsx
@@ -1,6 +1,16 @@
 import '../app/globals.css'
 import type { AppProps } from 'next/app'
+import { ThemeProvider } from '../components/theme-provider';
 
-export default function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+export default function App({ Component, pageProps }: AppProps) {
+  return (
+    <ThemeProvider
+      attribute="class"
+      defaultTheme="system"
+      enableSystem
+      disableTransitionOnChange
+    >
+      <Component {...pageProps} />
+    </ThemeProvider>
+  );
 }